requirements.txt
allure-pytest==2.13.5
pytest==8.3.3
playwright==1.48.0
pytest-xdist==3.6.1